requirements.txt
aniso8601==9.0.1
certifi==2021.5.30
click==8.0.1
cycler==0.10.0
Flask==2.0.1
Flask-RESTful==0.3.9
freetype-py==2.2.0
imageio==2.9.0
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
networkx==2.6.3
numpy==1.21.2
Pillow==8.3.2
pyglet==1.5.21
PyOpenGL==3.1.0
pyparsing==2.4.7
pyrender==0.1.45
python-dateutil==2.8.2
pytz==2021.1
PyWavelets==1.1.1
scipy==1.7.1
six==1.16.0
tifffile==2021.8.30
trimesh==3.9.29
Werkzeug==2.0.1
